+++ Content.IntegrationTests/_Carpmosia/Tests/MapRulesTest.cs
@@ -1,0 +1,150 @@
+#nullable enable
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using Content.IntegrationTests.Fixtures.Attributes;
+using Content.IntegrationTests.Fixtures;
+using Robust.Shared.ContentPack;
+using Robust.Shared.GameObjects;
+using Robust.Shared.Maths;
+using Robust.Shared.Prototypes;
+using Robust.Shared.Utility;
+using YamlDotNet.RepresentationModel;
+
+namespace Content.IntegrationTests.Tests;
+
+[TestFixture]
+public sealed partial class MapRulesTest : GameTest
+{
+    // Temporary override until most of the maps are fixed
+    private static readonly ResPath[] AllMapFiles = [
+        new("/Maps/_Carpmosia/Terminals/donk_rest_stop.yml"),
+        new("/Maps/_Carpmosia/amber.yml"),
+        // new("/Maps/_Carpmosia/bagel.yml"),
+        // new("/Maps/_Carpmosia/box.yml"),
+        new("/Maps/_Carpmosia/centcomm.yml"),
+        // new("/Maps/_Carpmosia/elkridge.yml"),
+        // new("/Maps/_Carpmosia/exo.yml"),
+        // new("/Maps/_Carpmosia/feint.yml"),
+        // new("/Maps/_Carpmosia/fland.yml"),
+        // new("/Maps/_Carpmosia/lampocteis.yml"),
+        // new("/Maps/_Carpmosia/marathon.yml"),
+        // new("/Maps/_Carpmosia/oasis.yml"),
+        // new("/Maps/_Carpmosia/packed.yml"),
+        // new("/Maps/_Carpmosia/plasma.yml"),
+        // new("/Maps/_Carpmosia/saltern.yml"),
+        // new("/Maps/_Carpmosia/snowball.yml"),
+        // new("/Maps/_Carpmosia/sparks.yml"),
+    ];
+    //private static readonly ResPath[] AllMapFiles = [.. GameDataScrounger.FilesInDirectoryInVfs("/Maps/_Carpmosia", "*.yml", true).Where(x => !x.ToString().StartsWith("/Maps/_Carpmosia/Legacy/"))];
+    //private static readonly ResPath[] StationMaps = [.. GameDataScrounger.FilesInDirectoryInVfs("/Maps/_Carpmosia", "*.yml", false).Where(x => !x.ToString().StartsWith("/Maps/_Carpmosia/centcomm.yml"))];
+
+    private const string PROTO = "proto";
+    private const string ENTITIES = "entities";
+
+    [SidedDependency(Side.Server)] private readonly IResourceManager _resMan = null!;
+
+    [Test]
+    [TestCaseSource(nameof(AllMapFiles))]
+    public void TestMapRules(ResPath map)
+    {
+        if (LoadMapYaml(map, _resMan) is not { } root)
+            return;
+
+        if (!root.TryGetNode<YamlSequenceNode>(ENTITIES, out var ents))
+            return;
+
+        List<string> errors = [
+            ..TestNonWallmountsUnderWalls(ents),
+            ..TestApcMissingConnections(ents),
+            ..TestPowerNetworkLabels(ents),
+            ..TestAnchorableDuplicates(ents),
+            ..TestUnlinkedAtmosDevices(ents),
+        ];
+
+        // Assert one large list of errors instead of Assert.Multiple to avoid 5 morbillion stacktraces
+        Assert.That(errors, Has.Count.EqualTo(0), $"Found {errors.Count} issues:\n{string.Join("\n", errors)}");
+    }
+
+    private static YamlMappingNode? LoadMapYaml(ResPath map, IResourceManager resMan)
+    {
+        var rootedPath = map.ToRootedPath();
+        if (!resMan.TryContentFileRead(rootedPath, out var fileStream))
+        {
+            Assert.Fail($"Map not found: {rootedPath}");
+            return null;
+        }
+
+        using var reader = new StreamReader(fileStream);
+        var yamlStream = new YamlStream();
+        yamlStream.Load(reader);
+
+        return (YamlMappingNode)yamlStream.Documents[0].RootNode;
+    }
+
+    private static YamlMappingNode? GetCompNode(YamlNode entNode, string comp)
+    {
+        var ent = (YamlMappingNode)entNode;
+
+        if (!ent.TryGetNode<YamlSequenceNode>("components", out var comps))
+            return null;
+
+        if (comps.FirstOrDefault(x => x["type"].AsString() == comp) is not YamlMappingNode trans)
+            return null;
+
+        return trans;
+    }
+
+    private static (EntityUid, (int, int), int)? GetApproxTransform(YamlNode entNode)
+    {
+        if (GetCompNode(entNode, "Transform") is not { } trans)
+            return null;
+
+        if (!trans.TryGetNode("parent", out var rawParent))
+            return null;
+
+        if (rawParent.ToString() == "invalid")
+            return null;
+
+        var parent = new EntityUid(rawParent.AsInt());
+
+        if (!trans.TryGetNode("pos", out var posRaw))
+            return null;
+
+        var rawPos = posRaw.AsString().Split(",").Select(float.Parse).ToArray();
+        var pos = ((int)Math.Floor(rawPos[0] * 10), (int)Math.Floor(rawPos[1] * 10));
+
+        var rot = 0;
+        if (trans.TryGetNode("rot", out var rotRaw))
+        {
+            rot = (int)Math.Round(MathHelper.RadiansToDegrees(double.Parse(rotRaw.AsString().Split(" rad").First())));
+        }
+
+        return (parent, pos, rot);
+    }
+
+    private static (EntityUid, (int, int))? GetTilePos(YamlNode entNode)
+    {
+        if (GetApproxTransform(entNode) is not { } trans)
+            return null;
+        var parent = trans.Item1;
+        var (px, py) = trans.Item2;
+        return (parent, ((int)Math.Floor(px / 10m), (int)Math.Floor(py / 10m)));
+    }
+
+    private List<EntProtoId> GetPrototypeIds<T>() where T : IComponent, new()
+    {
+        return [.. Pair.GetPrototypesWithComponent<T>().Select(x => x.Item1.ID)];
+    }
+
+    private static List<T> DeserializeCompNodes<T>(YamlSequenceNode entities, IEnumerable<EntProtoId> filter, Func<YamlNode, T?> deserializer) where T : struct
+    {
+        return [..entities
+            .Where(x => filter.Contains(x[PROTO].AsString()))
+            .SelectMany(x => ((YamlSequenceNode)x[ENTITIES])
+                .Select(deserializer)
+                .OfType<T>()
+            )
+        ];
+    }
+}
